Add Bazel build for message_generation (#66)
This also allows other Bazel projects (rosjava_core in particular) to
depend on this as an external repository.

It uses https://github.com/johnynek/bazel-deps to describe the Maven
dependencies. As a result, 3rdparty/ contains autogenerated files, and
the other files contain a hand-written description of the BUILD.

I will add documentation for the Bazel build to rosjava_core.